h2 and postgres.

+++ sinq-jpa/src/main/scala/io/sinq/func/Sum.scala
@@ -1,9 +1,7 @@
 package io.sinq.func
 
-import java.math.BigInteger
-
 import io.sinq.provider.Column
 
-case class Sum[T](val col: Column[T]) extends MethodColumn[BigInteger] {
+case class Sum[T](val col: Column[T]) extends MethodColumn[Long] {
   override def identifier(): String = "sum"
 }
